requirements.txt
streamlit
openai
pymongo
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
requests
PyGithub
python-dotenv
pytest